requirements.txt
Django==4.2.3
djangorestframework==3.14.0
rest_framework_simplejwt==0.0.2
